icepack_therm_vertical: allow initial 'zSin < min_salin' if 'l_brine = false'
In CICE-Consortium/CICE-svn-trunk@38a6bc4 (CICE 5.0 --- Beta version...,
2013-01-07), a check was added to init_vertical_profile (then in module
ice_therm_vertical) when initializing the melting temperature profile to
abort if the initial salinity is smaller than 'min_salin'.

The conditions of the check were then tweaked in
CICE-Consortium/CICE-svn-trunk@5f13d03 (Merging zbgc branch r606:675
into trunk... 2013-07-22) to only apply it if not using the mushy-layer
thermodynamics, and in CICE-Consortium/CICE-svn-trunk@32abea3 (Changes
to get 0-layer model working again... 2013-08-02) to restrict the check
to the BL99 thermo.

With the BL99 thermo, one can disable brine pocket treatment by setting
'saltmax' smaller than 'min_salin', which in turns sets 'l_brine' to
false in 'icepack_init_thermo' and initializes the salinity to zero.
In that situation it does not makes sense for init_vertical_profile to
abort since the salinity will always be zero and thus lower than
min_salin.

As such, tweak the conditions of the check to only apply it if 'l_brine' is
true.