icepack_atmo: correctly compute Tref, Qref at zlvs
When we compute the diagnostic temperature and humidity (at zTrf) at the
end of icepack_atmo::atmo_boundary_layer, we shift the stability
parameter (hols) by multiplying it by zTrf/zlvl, but this is incorrect
when staggered atmospheric levels are used, as zlvl is the atmospheric
level height for momentum, not for scalars (temperature and humidity).

This is a bug introduced in 37f2a17 (Add support for staggered
atmospheric levels (#364), 2021-05-28) but only changes diagnostics
if staggered atmospheric levels are used, i.e. if zlvs /= zlvl.

Correct that by dividing by zlvs, the atmospheric level height for
scalars, when it is present.

While at it, adjust the indentation of the description comment for
'zlvs', so it is aligned with the other variables.