icepack_therm_vertical: rename worka,workb to delt,delq
In icepack_therm_vertical::icepack_step_therm1, the local variables
'worka' and 'workb' are sent to icepack_atmo::icepack_atm_boundary as
the 'delt' and 'delq' arguments. This is the sole use of these variables
in icepack_step_therm1.

These names date back to the use of global, temporary work arrays in
CICE4. Even if their values are not used, let's make the code clearer by
renaming the local variables delt and delq.

Closes https://github.com/CICE-Consortium/Icepack/issues/375