squash! icepack_parameters: make 'zTrf' a parameter
In cd268d5 (icepack_parameters: make 'zTrf' a parameter, 2022-03-25),
the "Compute diagnostics" comment at the end of
'icepack_atmo::atmo_boundary_layer' was changed to mention 'zTrf'
instead of hard-coding 2 meters, but at the time I failed to notice that
this comment was erroneous as the diagnostic wind speed is computed at
zref, not at zTrf.

Fix the comment now.

This commit is titled so that a 'git rebase -i --autosquash' will fold
it into its parent when we rebase its containing branch before
submitting an upstream PR.