icepack_therm_vertical::icepack_step_therm1: optionnally return surface specific humidity
In a coupled context, calling models might be interested in the surface
specific humidity computed in icepack_atmo::atmo_boundary_layer (stored
in the local variable 'ssq'). This variable is not returned by that
subroutine, but 'delq', the humidity difference between the surface and
the atmosphere, is. Since we have (from atmo_boundary_layer)

     delq(i,j)  = Qa(i,j) - ssq

we can easily recover the surface humidity from the difference:

    ssq(i,j) = Qa(i,j) - delq(i,j)

Make that quantity available to calling models by adding a new optional
'intent(out)' argument to icepack_step_therm1, 'Qsur' (for "surface").

Since the call to 'atmo_boundary_layer' is inside the loop on
categories, add a new local variable 'delqn', send it to
atmo_boundary_layer instead of 'delq', and aggregate the value over
categories in 'merge_fluxes', making sure to initialize delq _outside_
the loop. Compute Qsur from the aggregated delq at the end of
icepack_step_therm1.

Co-authored-by: Frederic Dupont <frederic.dupont@canada.ca>