squash! icepack_parameters: make 'zTrf' a parameter
In the previous commit, we fixed the comment in atmo_boundary layer to
account for the fact that Uref is computed at zref, not zTrf, but we
failed to correct the describing comments added in icepack_parameters
when 'zTrf' was added. Fix this now by only mentioning Tref and Qref.

This commit is titled so that a 'git rebase -i --autosquash' will fold
it into the appropriate commit when we rebase its containing branch
before submitting an upstream PR.