icepack_flux: remove 'optional' qualifier from Uref argument in merge_fluxes
When the computation of Uref (in ice_atmo::atmo_boundary_layer) was
added in CICE-Consortium/CICE-svn-trunk@7f79790 (Merged changes between
trunk r779 (version 5.0.4) and cesm_cice branch r814 into trunk.,
2014-08-26) [1], that variable was added as an optional argument. A
corresponding optional argument was thus also added in
ice_flux::merge_fluxes to aggregate that value over categories. The
choice to use optional arguments is not explained in the commit message,
but we can guess that it was to avoid potential external callers of
those subroutines (as in, external to the CICE repository) having to
add that additional field if they were not interested in it.

The optional qualifier was kept in merge_fluxes when the column physics
code was extracted into the Icepack repository (this happened during
early development of the 'colpkg' and is not recorded in the Git
history), even if the sole caller of merge_fluxes, icepack_step_therm1,
has 'Uref' as a non-optional argument. This means that there is no
reason now for this argument to be optional. 'merge_fluxes' is not part
of the Icepack API, so let's make its 'Uref' argument non-optional.

This incidentally corrects the location of the "ocean fluxes" comment,
which is misplaced as 'Uref' is not an ocean flux, rather an atmospheric
flux.  It has been misplaced ever since Icepack's initial commit 6f98a36
(initial commit of icepack to git from LANL consortium branch r1229 May
24, 2017, 2017-05-24).

It is at the right place in last commit imported from SVN on the
'master' branch of the pre-CICE6 repository,
CICE-Consortium/CICE-svn-trunk@ac5d481 (trunk: Merge r1132:1133 from
rasm_colpkg branch..., 2016-07-29) [2], so this must also have happened
during early developpement of the 'colpkg'.

[1] https://github.com/CICE-Consortium/CICE-svn-trunk/commit/7f79790588a74ff4e48b23dee1db0702727d5d44
[2] https://github.com/CICE-Consortium/CICE-svn-trunk/commit/ac5d48162210ac021dfa2c0d9e2c43a011a84a73