icepack_therm_vertical: mention failing category when thermo_vertical fails
In icepack_step_therm1, we check if the call to 'thermo_vertical' was
successful and add "Vertical thermo error" to the Icepack warnings if
not. This is printed after any warnings added in 'thermo_vertical'.

If the temperature solver failed, it can be useful to know which ice
category is the failing one. This can't be outputed in thermo_vertical
since the loop on categories is in icepack_step_therm1; thermo_vertical
does not know which category it is solving for.

Add the failing category to the Icepack warnings if 'thermo_vertical'
resulted in an abort.

A quick audit of icepack_step_therm1 reveals the only other subroutine
that potentially aborts the model for physics reasons is
compute_ponds_topo. Other subroutines do call 'icepack_warnings_aborted'
under some conditions, but these cases are when subroutine arguments are
inconsistent (see for example atmo_boundary_layer).  Since
compute_ponds_topo is called outside the loop on categories, we thus do
not need to add the failing category anywhere else in
icepack_step_therm1.