icepack_therm_shared: increase ferrmax to 2.0E-3
The parameter 'ferrmax' in module icepack_therm_shared represents the
maximum allowed energy flux error in both temperature solvers (BL99 and
mushy). Its hardcoded value is 1.0E-3 but it was increased to 2.0E-3 in
our in-house version of CICE 4 and 5, following cases where the solver
failed to converge due to this value being too strict.

Such a case was encountered during the initial tests of CICE6 with
Delta-Eddington in the GDPS, see [1] and following comments.

Change 'ferrmax' to the value we were using in CICE4. Note that this is
sill 5 times smaller than the recommended value (see code comment).

[1] https://gitlab.science.gc.ca/atmospheric-data-assimilation/gdps/issues/415#note_632120