icepack_therm_vertical: correctly compute surface specific humidity (Qsur)
In 295b6b8 (icepack_therm_vertical::icepack_step_therm1: optionnally
return surface specific humidity, 2023-03-22) we let
icepack_therm_vertical::icepack_step_therm1 return 'Qsur' as an optional
output argument. As explained in that commit's message, this is
implemented by aggregating over categories the value of 'delq', the
humidity difference, and substracting it to 'Qa', the reference level
specific humidity, to recover 'Qsur' at the end of the subroutine.

This approach is wrong and leads to (widly) erroneous values whenever
the ice concentration is not equal (or very close) to 1.0, i.e. at the
ice edge. The reason is that we omit to multiply the reference humidity
by the ice concentration. The correct computation is:

    Qsur = \sum_n Qsur^n A^n
         = \sum_n (Qa - delqn^n) A^n
         = \sum_n Qa A^n - \sum^n delq^n A^n
         = \sum_n Qa A^n - delq

Notice how 295b6b8 instead wrongly computes

    Qsur = Qa - delq

omitting the factor \sum_n A^n = aice.

Fix that by aggregating Qsur directly instead of aggregating delq. Add a
local 'Qsurn' variable in icepack_step_therm1, compute Qsurn immediately
after the call to icepack_atm_boundary, and send Qsurn and Qsur to
merge_fluxes. As such, rename the argument 'delq' and 'delqn' to 'Qsur'
and 'Qsurn' in merge_fluxes.