icepack_atmo: compute Uref in the same way as [TQ]ref
In CICE-Consortium/CICE-svn-trunk@7f79790 (Merged changes between trunk
r779 (version 5.0.4) and cesm_cice branch r814 into trunk., 2014-08-26),
which went into CICE 5.0.4, subroutine ice_atmo::atmo_boundary_layer was
changed to also compute the module of the diagnostic wind (Uref).

The approach used is different than what is done for diagnostic
temperature and humidity: we simply multiply 'vmag' ('wind' capped at
'umin') and 'rd/rdn' to shift the value to the diagnostic level (in the
'highfreq' case, the relative difference between wind and ice velocity
magnitudes is used instead of 'vmag'). In contrast, for temperature and
humidity we shift the value of 'hols', recompute the stability function
at the reference level and compute the diagnostic values by adjusting
the input variables 'potT' and 'Qa' using the differences 'del[tq]' and
a 'fac' factor computed from these stability functions.

In our in-house version of CICE4, we also compute 'Uref', but using the
strategy just described for 'Tref' and 'Qref' [1]. To minimize changes
with our in-house version of CICE4, change the computation of 'Uref' to
use the same strategy as 'Tref' and 'Qref'. Note that this leads to
relative differences for 'Uref' of the order of 50 %, which is
considerable.

This is most likely because when 7f79790 introduced the computation of
'Uref', it was in the context of the RASM or CESM systems where the
number of iteration (hard-coded at the time, made into a namelist
variable in CICE-Consortium/CICE-svn-trunk@192f1eb (Merging r825:842
from cesm_cice branch into trunk. ..., 2014-10-02)) was higher than then
default value of 5. My understanding (this would have to be verified) is
that as we increase the number of iterations, both approaches give more
and more similar results.

Finally, note that we do not change the computation under
'highfreq=.true.', for now, since this option is not used in our
systems.

[1] https://gitlab.science.gc.ca/concepts/concepts/blob/master/CICECMC/cice4.0_cmc/source/ice_atmo.F90#L471-486