icepack_therm_shared: add tfrz_option='nonlin' in icepack_liquidus_temperature
In CICE commit CICE-Consortium/CICE-svn-trunk@9276469 (Removed namelist
option Tfrzpt..., 2013-08-20), which went into CICE 5.0.0, the 'Tfrzpt'
option was removed. Before that commit, ice_flux::init_coupler_flux
initialized 'Tf', the ocean freezing temperature, to a constant ('Tfrzpt
=constant') or as a linear function of salinity ('Tfrzpt=linear_S').  In
that commit this initialization was changed so that Tf was initialized
based on 'ktherm' instead, with an equivalent computation as
'liquidus_temperature_mush' for mushy-layer thermodynamics, and as a
linear function of salinity for the BL99 thermo. Subroutines in
'ice_forcing' relating to ocean forcing were adjusted in the same way.

Then in CICE-Consortium/CICE-svn-trunk@f6f27b5 (Merging cesm_cice
branch r896:924 to trunk, ..., 2015-03-04), which went into CICE 5.1, an
almost equivalent option 'tfrz_option' was introduced. This option was
used in the different forcing subroutines via a new
'ice_forcing::ocn_freezing_temperature' subroutine used to compute the
freezing temperature based on the ocean forcing, accepted 'mushy' and
'linear_salt', and defaulted to (the equivalent of) 'constant'. Note
that ice_flux::init_coupler_flux was _not_ adjusted at that time to use
this new option, so the 'Tf' initialization was left as-is, i.e.
depending on 'ktherm'.

This lead to the introduction of two similar subroutines in Icepack:
icepack_liquidus_temperature, which is used in CICE in
ice_flux::init_coupler_flux, and icepack_sea_freezing_temperature, which
is used in ice_forcing::ocn_freezing_temperature.

The commit message of f6f27b5 includes this note: "the optional values
for Tf need to match what is done in a coupled ocean model". Since NEMO
uses yet another different formulation for the freezing temperature, our
in-house version of CICE4 allowed a new value for 'Tfrzpt', 'nonlin',
corresponding to the computation done in NEMO in 'eosbn2::eos_fzp' [1].

To bring the code closest to our in-house CICE4 version, adjust
icepack_liquidus_temperature to initialize 'Tf' based on 'tfrz_option',
as was done in CICE4, and allow 'nonlin', copying the computation from
NEMO's 'eosbn2::eos_fzp'. This is done using the CICE_IN_NEMO CPP macro
to minimize impacts on other models. Note that we can't call the
'eos_fzp' subroutine directly since this creates circular build
dependencies, which is why we copy the computation. Note also that this
subroutine also allows a different computation based on NEMO namelist
option 'nn_eos', but this is not taken into account here ('nn_eos=0' is
assumed).

Finally, note that in retrospect, this change is probably not needed at
all when running _coupled_ to NEMO, since in CICE the salinity used to
initialize Tf is the hard-coded constant value in
ice_flux::init_coupler_flux, not the one from NEMO, which is part of the
reason 'ice_init::init_state' is called a second time in
sbcice_cice::cice_sbc_init. When running standalone with NEMO forcing,
this commit would probably make sense, but then this is not something we
do anymore, and we would need equivalent changes in
ice_forcing::ocn_freezing_temperature.

[1] https://gitlab.science.gc.ca/concepts/concepts/blob/master/NEMOGCM/NEMO/OPA_SRC/TRA/eosbn2.F90#L1029-1036