icepack_therm_shared: avoid division by zero when initializing over fresh water
When Tf is computed with 'tfrz_option=nonlin', it is possible that some
cells have Tf=0, if the SSS is zero (for example, we have seen this in
the Baltic sea).

This leads to a possible division by zero when initializing the ice
temperature profile (under 'calc_Tsfc=.true.') since we have

    Tsfc = min(Tsmelt, Tair - Tffresh)
    slope = Tf - Tsfc
    Ti = Tsfc + slope * (k-p5) / nilyr

When Tair is above zero degrees Celsius, we have Tsfc = Tsmelt = 0, Tf =
0, and so slope = 0 and Ti = 0. We then divide by Ti when computing the
enthalpy of each layer.

Avoid this division by zero by simply computing the enthalpy of fresh
ice in this case (this is the same computation as is done for the snow).

Reported-by: Ryan Muncaster <ryan.muncaster@canada.ca>
Co-authored-by: Francois Roy <francois.roy3@canada.ca>