icepack_atmo: compute Jordan stability function in psi_stable_jordan
The computation of the Jordan stability function, used for the stable
case, is currently done in subroutine compute_stability_function,
whereas the computations for the unstable case are done in their own
functions, psi_scalar_unstable and psi_momentum_unstable.

Move the computation of the (stable) Jordan stability function to its
own function, psi_stable_jordan, and call it in
compute_stability_function. This will allow us to call psi_stable_jordan
directly in a following commit.